Tableau dashboard stripped of specific information and data.
This is a dashboard created to monitor Inventory Control related key metrics. Trends of high level "shrink" of inventory detailed. Breakdowns of that data follow by certain groupings and can be customized by different elements.

All elements are generic to avoid any potential breach of sensitive information. All synthetic data was generated randomly and is not related to real information in any way.